Selenium fixture has better support for navigation fixtures
Given that the selenium fixture is only being used in its tests,
I thought now would be a good time to improve its interface to be
more useful. It can still navigate you to arbitrary URLs, but once
there it is contrained to using a navigation fixture. This allows
for a few assumptions to be made that drastically improve this
fixture's usability.