events.html fix, lint
Professional blindness, sorry please :). I've accidentally set the logging to point to the resulting html file. As the EventListener object gets created every time even if no event testing is active due to the fixture that is provided, it always created the events.html file. test_events.log is always created, because of this, but this file does not matter, it's for debugging purposes.

Also applied some lint for issue that arised when I fixed the ipecho problem.