Move examples into their respective docstrings
a.k.a. "The one that psav might not like"

Having separate sections for the examples, particulary where the example
wasn't on the same page as the code documentation (accordion, toolbar),
isn't as handy as having the usage right next to the signature, and
Args/Returns/Raises blocks. This was made comical by using the
sphinx-apidoc generator.