Tweak the third: Table constructor is easier to work with for normal tables
Many (most?) tables in CFME are "normal" tables, with thead and tbody children.
Table now assumes it's dealing with such a table. In cases where the tables are
abnormal, and the header body rows are in different tables, SplitTable joins the
party to handle the ridiculousness.

This commit also includes additional docs.