quadicon locate always returns anchor
We had two locators for quadicons:
- '9.9.9.9': "//div[@id='quadicon']/../../..//a[@title='%s']" % self._name,
- 'default': "//div[@id='quadicon' and ../../..//a[@title='%s']]" % self._name

9.9.9.9 returns the anchor element, default returns div#quadicon.

My 5.2.3 appliance was breaking because it was getting the div back, and
doing sel.click on the quadicon div wasn't triggering navigation to the
quadicon div's anchor href. The anchor found by the 9.9.9.9 locator is broken
on downstream, but the div anchor seems solid on both.