Select doesn't wrap other Selects
Moves deferred Select type into pytest_selenium

We were using the selenium Select class in pytest_selenium, but one of the
main goals of that plugin is to be able to instantiate locators without having
a browser open. Therefore, our Select had to move from web_ui to pytest_selenium.

See Select's __init__ method in the before/after for the differences in handling
`self._loc`