Fix linear nav mouse move issue
In the navigation selenuim seems to move from point A to point B in a
linear vector. Consequently in the navigation if this vector passes over
another top level navigation menu, the top level will be selected and
the intended second level navigation item will become invisibile leading
to a failure. This fix opens the top level navigation and then drops to
the first second level element, before moving to the intended second
level element, effectively creating an L shape two step navigation,
instead of a straight diagonal vector.