Fixed testgen to generate the right names/values
Argnames and argvalues would get out of sync because fields lookup tried
searching for fields names in the yaml as well as using as a special
arg. This would offset the argnames and argvalues, leading to severe
problems. This only occured if a user requests a special arg, or if the
user requests specialargs out of order from the sepcial arg typemap.