Always take a browser screenshot
Many tests are failing because they aren't marked to use the browser,
despite doing so in the course of their tests.

So, rather than inspecting the fixtures to decide whether or not to take
a screenshot, we'll just go ahead and take one. There's enough info in
the artifactor report now that we can easily decide if the screenshot is
useful.