Toolbar sub_loc does an exact text match
This deals with the issue of two very similar
sub menu items resulting in choosing the wrong item.

The specific example that requires this change is two sub menu items, in this order:
- Add a new Button Group
- Add a new Button

"Add a new Button Group" appears first in the DOM, and would always have been
selected, despite "Add a new Button" being the expected choice.

This will break toolbar selects that relied on the previous substring
matching behavior, but being explicit in those cases seems like the better option.