Template tracker tools
Trackerbot will be handling more than just template, but for now the
populate script is all we've got. Soon we'll need some scripts for jenkins
to mark templates as tested/usable, and a simple tool for getting the
latest template (and its providers) for a given stream.

Soon: Actually deploy the trackerbot and drop the api settings into env.yaml