removed assets check in precompile_assets
jweiss discovered a case where the assets dir can exist but the assets
still need to be precompiled, so for now I've removed the check, and we'll
need to revisit this to find a better way to see if assets need to be precompiled or not.

Maybe keeping the check and adding a --force flag is the right answer. I'm not sure...