Fixed a DB query that was breaking nav
This is a workaround, and not a fix for the actual problem.

This only affects openstack instances, and is caused by the server filter
matching appliances based on IP addresses. Openstack instances usually have
two IP addresses, where one is private to openstack and the other is a routable
floating IP address. The appliances store the private address, so the address
that we're matching against in the miq_servers table doesn't match our base_url,
and the server info can't be found.