Fixed filedumps to use correct artifact location
Filedumps in the artifactor_plugin were still using the node name
and parent name, where as the rest of the artifacts were stored using
the location data from the node. This update fixes that and makes the
browser.py calls using the filedump artifactor plugin switch to using
the right test name.