Fix datastore listing on vsphere
vmware datastore listing was recently changed to ignore inaccessible
datastores. In fact this was not the problem, CFME does not show up a
datastore on a provider if it is not mounted by any hosts.

This commit changes the datastore listing function to mimic that
behaviour by checking if the datastore is mounted on any hosts.