Fix template not found issue
Issue was that the templates headers changed from cloud to infra
the 'tenant' column is added in cloud. This gave an off by one error
where the internal header cache of the table didn't match what was on
the page. This has been fixed with an update cache call inside the
navigatation function. The fix may be supplanted by something better in
the coming days, but this should get things working again.