Menu fixes
Menus have changed to a new style so in order to support them we need to
version pick the locators. Since we don't want to do that at import
time we create a dumb class to defer the version picking as a property
attribute. Future work would be to ensure that as much version picking
is done at the top level. However, there is due a new menu module which
will reduce the navigation overhead considerably.