py.test collection works with nonexistent cfme_data
Most of the changes were made witha  gnarly sed line, but there were
some uncollect improvements to be made in a few places. There is much more
that we can do on the uncollecting front, but this gets rid of all collection
errors related to missing keys in cfme_data.