Parallelizer doesn't block on slave shutdown, has better logging
The master will wait a polite amount of time for a slave to shut down,
set at 10 minutes for now. If more time than that goes by, the slave is
killed. Additionally, py.test will log not only that a slave is shutting down,
but that it has, in fact, shut down (or died).

I also got to remove some vestigial event ordering functionality that became
useless when we got everything parallel happening in master's runtestloop.