Fix element() masking the force_navigate effort
By introducing the check for rails errors in element(), force_navigate's efforts to kill browser were not possible to do because it detected the error in is_displayed() (which uses element()) and bailed out. Short term solution is to add _no_deeper=True to the function call, long term one is to unify and DRY the rails detection code.