Protect against other exception types
If the exception is not one that we explicitly generate then the second
argument may be missing. This caused issues in some runs and so to
mitigate we now check the length of the Exception arg to be 2 before
trying to access the port. If it's not present we just set port to None,
which doesn't harm anything.