Sprout - Care about template age.
Groups can now be configured with an 'age' of template to be considered obsolete. If the template is older than that number of days, it won't be pulled from trackerbot. If the option for deleting is on, it will also delete such templates if they are not the latest (therefore the only) ones.