Force checking of ports in appliance policer
Funny thing, if you don't erase the evidence, the Appliance Police
will keep on arresting you. :( This fix forces the Appliance Police to
make a new check of the port availability everytime. We hadn't seen this
issue crop up before because most failures had been happening due to the
UI worker dying. This meant that 443 was still contactable and hence the
cache that the utils.net keeps, did not affect the Police.

If a port is uncontactable, because of network issues for example, then
the fact the port is unreachable was cached in the utils.net module and
because the force flag was absent was never rechecked and the cached
result just served up again and again. Thus continuing slaves just led
to them getting pulled over by the Appliance Police again for crimes
they hadn't committed.