zmq unsnarled
The "snarl" referenced in the removed comment was being caused by
the recv thread reading the socket while the main thread was
simultaneously writing to it. As such, while the sleep would make
the problem less likely, it wasn't a real "fix" (when has sleep
actually fixed a problem?). Putting socket reads and writes under
a thread lock seems to do the trick.