Fixed another collections related issue
In jenkins runs we were seeing rhos_uc being used as a valid provider,
but receiving the scvmm management object. This was because it was not
being overwritten in the loop with a blank object. The error was
manifesting because the type checking had no failsafe. This has now been
corrected and the provider is skipped if the type is not valid.