Sprout - Correctly delete pool after salvaging appliances
The code that deletes the pool when pool is killed is placed in Appliance's .delete() method. That ensures that when the last appliance in the pool gets deleted, the pool gets deleted with it. However when we kill an unfinished pool and the appliances are salvaged, this obviously does not happen when the appliances are put in shepherd. So an additional check is required.