Able to override BZ stream detection in blocker
If you want to make the bug skip for stream that is not cloned in BZ, you can use keyword parameter forced_streams=[] which takes the stream names in form of downstream-54z and so. It will then always care about htat bug in that stream no matter what do the bug's versions state. Use with caution though...