Resolve the appliance configuration error
What was the issue? The problem was, that somewhere in the configuration function, a call to version.pick was done. Since the appliance object that we are calling it from is not pushed to the appliance stack, it actually runs the version.pick against some other appliance. The whole fix is to enclose the configure sequences with `with self.ipapp` that ensures version.pick uses the correct address.