If the element is not movable to, text() won't fail
Some of the elements we want to read text from are not properly movable to, like with the new selects. This is a workaround, that when it cannot scroll to the element, uses JavaScript to read it. I believe that JavaScript solution is much faster, however I also believe the approach with move_to_element should be the primary one as that is what the user does.