Fix the is_dimmed() function in paginator. And regexps to parse numbers.
Last modification of the paginator locators changed the target element from the 'li' to the one-level deeper 'span'. is_dimmed() checked the class attribute of the element for presence of the dimmed/disabled class, but that is only in the 'li', not in 'span', so I added a check that if it is a span, go one level up and check it there.

There has also been a change in the format of the paginator text, I changed the regexps to parse them too.

{{pytest: -k visual -v --long-running}}