Work around the importerror with yaml files
When there are both yaml and eyaml files, it tries to print out a warning that it ignores the encrypted yaml during the import. However, as this is happening during the import, the module with the yaml is not loaded yet and the warning print gets delegated to artifactor, which will then also try loading the same config file (credentials), resulting in import error.