Fix toolbar name collision
A strange situation occured where the locator for the sub actually
matched the main link element for the Timelines page. Subsequently when
we were trying to select the timelines toolbar for monitoring, the first
element being returned was actually the element from hte main nav, which
didn't have the 'idd' attribute. So we now see if multiple elements were
found, which one has the idd element.