Fix ssh_client cache invalidation
This is a quick fix to invalidate the ssh_client cache instead of
setting the attribute to None. Setting to None does not allow the
ssh_client to retry the operation with the new port. This was probably
leftover from an earlier appliance makeup.

Proposal that we go through and actually make this fire a signal to
invalidate the proper cache settings, rather than just the SSH, ie, the
DB would probably need invalidating too to make things _right_.