Add seal argument in templatize method, tweak appliance script
I need to opt out with the sealing when using inside CFME because I do it in a different state machine. This also requires me to add some code that processes some known strings into actual values, like boolean. I changed it to run for booleans and none too. This code should be later extracted out into some tools if we find use for that. Added a small unit test for the argument processor.