Try fixing artifactor NOT logging things properly
When the formatter change occured, all the log calls were updated to be "conforming to the norm". However, the artifactor logger was optimized for the old way, and it was optimized so much it did not digest the *args at all. This tries to rectify the issue by doing the formatting on the client side. If we ever decide to do something special with the params inside artifactor, this will need to be changed. But I hope it won't be.