Table - Handle hiding and some additional tweaks
Handle hiding of the table. Involves the .rows() and the things that use it directly. The other things will eventually need this change too once needed.

I also added an exception when you click a row and it is not found.

Removed the convert_header in favour of attributize_string from the utils which does the same.

Fixups in some places in tables.