Fixed menu lookup error
This was an awesomely weird and bad bug. Owing to the deferred nature of
the menu instantiation, now happening inside force_navigate(), the
branch stack processing actually happens inside a test context. This
meant that if a LookupError happened inside the menu initialization(),
it would raise the exception, be caught as a test failure, and the
framework would continue testing, unaware that the branch stack hadn't
necessarily finished being processed. This change detects a LookupError,
and logs a warning. We also raise in the case of any other failures.