Fix testgen mark application
Previously the mark was applied correctly, however in the testgen
refactor this became broken. As the mark was applied outside of the
loop, merely running "through" testgen was enough to mark the
test. Unfortunately this was not accurate. This mark has now been moved
to be applied only if the "provider" parameter has actually been used
and hence the test has been properly parametrized.