Parallelizer: Ensure the timer is always rescheudled
A rogue exception could have caused the problems we had. When I wrote this, I forgot to put the reset timer function in the finally block. Since an unexpected non-sprout exception is something potentially dangerous, doing a short timeout in that case to "wait" for the good state to come back..