WIP: Fix timeouting wharf browser.
The wharf browser session was timeouting and closing abruptly with
URLError or Bad status line because the renew mechanism was flawed.

When the Timer(subclass of threading.thread) triggers, the Timer.is_alive()
returns True. There was a check that caused skipping the renewal in that case.
The renewal probably never worked up until now.