ipappliance nonzero exit fix. (#3926)
Ensure that we getting nonzero exit from the process for any problem in
call_appliance() -- for example if the IPAppliance instantiantion fails,
it used to return 0 because an Exceptin there meant just thread exit,
without writing to the results.

Threads can be very cumbersome whereas futures are quite elegant way to
spin-up tasks on multiple threads. Let's use it to to make code hopefully more
readable and less error-prone.