Temporary move to current_appliance
When we finally move to collections on these objects, it will be
possible to pass the db through to them from the self object inside the
collection. Until then, we will use current_appliance. It's no worse
than using cfmedb() because that used store.current_appliance in the
background anyway.