Snapshots tests adoption for RHV.
Hi!

I've took https://github.com/ManageIQ/integration_tests/pull/4562
pull request and it wasn't working properly. As far as I understood
it's because of RHV VM's quadicon state could be not just
'currentstate-on' but 'currentstate-on-<some_hash_here>'.

So I've changed soft assertion's condition and now it's fine.

Thanks.

Signed-off-by: Aleksei Slaikovskii <aslaikov@redhat.com>