[1LP][RFR] Automation support for Openshift pods testing. (#3679)
* Automation support for Openshift pods testing.

This is what I figured out so far. Works similarly like the docker container solution but this one uses a parent ssh client which points to the host machine. The pod name is specified as the already existing container field in env.yaml. New field (openshift) added to env.yaml which shall contain hostname/ip, username and password so an SSH client can be created pointing to the openshift host machine in order to execute the commands on the pod. It uses the oc rsh command to execute the command in the pod.

* Rework the SSH system for openshift

* Database with openshift lies on a different host

* DB host override support

* Openshift pod file upload

* File downloda working

* Some logging updates

* Fix logging calls to conform the standard

* Fix IPAppliance signature doc

* Fix SSHClient signature doc

* Additional logging calls for file transfer

* Use check_call instead of call

* Remoe the double newline