Superspeed
dismiss_any_alerts takes ~2s and is invoked on every step. By putting it inside the following try/except we only execute it when there is a possibility it actually happened. If there is an alert and you try to interact witha page, you get an exception, which we handle, currently we do nothing but this patch moves the alert check in there.