Add scripts/list_provider_vms
This simple script is for monitoring of active vms on providers.
Takes --tag and positional provider key options
tags correspond to those in cfme_data for management_system entries
Multiple tags are additive, not combination
Can specify tags and providers

Takes --outfile to write a tabulated report to a file, also writes it to
stdout along with progress updates

Lots of other things this _could_ do, but keeping it simple for now.