Added new BaseCollection and BaseInstance classes
Currently we have a need to differentiate between instances and
classes, where as before they were both "identical" in a sense that
you could just use them as classes without instantiation, during this
transition period it's not the case. All the current Collections
should _very shortly_ no longer work in this way and we will need to
instantiate a 'collection' object.

These new classes allow isinstance usage and also for better support
further down the line with regards to ensuring API.