Use paramiko recv_ready as function
Cause it is a function not a property.   This happening in ssh.py's
run_command() was causing some programs to hang when ran.
Essentially when the recv_ready function for that file was checked
without the parenthesis, it would return an object, and object is always true
in the context of an if (well unless operator overloading in boolean context
was used). Anyway, needed to add () at the end of those calls.

I'm actually not sure why it wasn't seen with just any program,
but maybe somehow short running programs did not experience the
problem.   But the end result was if you had a either stdout or
stderr empty, it would still try to read from it and then the
paramiko read() code would go into a loop where it slept till
timeout.