Make coverage_merger.rb not replicate file data anymore
When we merged the coverage data we would replicate all
.resultset.json files in one file, merge/.resultset.json. This
file in some cases would be 100Mb's. These changes make it
actually merge the file coverage data for reach file, thus
not having replicated data. Merging the data in this way makes
the file like 2Mb in size, and makes the sonar-scanner run for
less than 2 minutes (as opposed to approximately 50 minutes).