Turn eventing ON for Nuage provider
With this commit we extend Nuage provider so that it now inputs
AMQP credentials upon creation. Until now we were only inputing
the Default credentials.

Also, with this commit we update test that was only testing
- CREATE
- RETRIEVE
- DELETE
of Nuage provider to now also validate
- UPDATE
which gives us a complete CRUD test in the end.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>