Add ability to aggregate coverage from multiple jenkins jobs
In addition to aggregating multiple Jenkins jobs I had to do a
work around because we ran out of disk space on the appliance.
The work around was to extract the coverage tarballs in waves,
merge that data, cleanup and use the merged data in the next wave.
Taking this route allowed me to quickly get this working, but over
time I think we seriously need to look at a different way of storing
and getting out the coverage data from test runs.