[1LP][RFR] Add ability to aggregate coverage from multiple jenkins jobs (#6916)
* Add ability to aggregate coverage from multiple jenkins jobs

In addition to aggregating multiple Jenkins jobs I had to do a
work around because we ran out of disk space on the appliance.
The work around was to extract the coverage tarballs in waves,
merge that data, cleanup and use the merged data in the next wave.
Taking this route allowed me to quickly get this working, but over
time I think we seriously need to look at a different way of storing
and getting out the coverage data from test runs.

* Code review #1 PR 6916

The particular issue in question was the proper use of namedtuple
classes.   Sorted this out, so I wasn't creating a Build class over
and over again.   Also, had to deal with the fact that namedtuple
generated classes are immutable.

* Move check for no eligible jobs to account for multiple jobs.

* Code Review #2 for PR 6916

Calls to ssh.run_command have been wrapped to ensure consistent
logging and error handling.