Ensure mostly same files show up in coverage data between test runs
Made coverage_merger.rb add coverage data for files that were not
loaded at all but are considered part of the source.   This should
give us more accurate coverage statistics when we run tests that don't
include some of these files.

To create the coverage data simplecov's LineClassifier is used from
version 0.16.1 of simplecov.  Since it won't install in the CFME
gems dir, I had to install in the system gems dir, and make
code changes associated with that.