Temporary hack track coverage in forked CFME processes
We are losing coverage statistics because the way we
setup CFME to give us that data was designed around its
2014 architecture where it would launch multiple processes.
It changed at some point to launch a single process that forked.
What is happening now is there is we are only getting the statistics
from the parent and not from the children.

This is a temporary hack, to get this info until this PR is submitted
and makes it into a build:

   https://github.com/ManageIQ/manageiq/pull/17302

And even that is a temporary solution until CFME takes ownership
of generating code coverage statistics as a full feature.