add switch to summary for physical infra provider (#7629)
In version 5.10, the Details view for a physical infrastructure
now has a choice between a Dashboard View, which is new, and
a Summary View, which has been refreshed but it contains elements
of the previous Details view. By default, the Dashboard View
is opened, so we added a step to select the Summary View before
doing the stats check.