Change sprout client default auth behavior
Remove default credentials lookup for sprout client, using only env vars or a passed creds key

This prevents groups with a credentials file that contains a 'sprout' key from defaulting to that user for anyone using the credentials file.
This was causing regular users to make sprout requests as a CI user instead of themselves.

Bad tuple unpacking was setting pool attribute to tuple of IPAppliance.from_json call and the pool ID.