Don't crash fixture if Merkyl plugin returns bad response
With this commit we prevent crash of entire `merkyl_inspector` fixture
in case Merkyl plugin timeouts. Thing is that Merkyl service on appliance
sometimes timeouts before replying to our requests. And currently we
crash hard when this occurs. But we shouldn't - because Merkyl service
usually returns valid response upon next probe.

So with this commit we simulate empty response as opposed to crashing.
The real fix, obviously, would be to fix merkyl service itself, but that's
out of our current scope.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>