Support reset log for merkyl plugin
With this commit we implement method that resets log for
merkyl plugin. Now we can clear content of a log and
continue tailing it.

This is needed when a log is tailed for a duration of the
whole session and there are multiple tests being ran. For
example two test can search for the same log entry. If
first one succeeds and we don't clear the log content
after, second one will just find the log content of the
first test and succeed.

Signed-off-by: sasoc <saso.cvitkovic@gmail.com>