Move Nuage fixtures
With this commit we move two of Nuage fixtures to fixtures folder
and register them. Now they are now available in global fixture list
and can be used by multiple tests.

`with_nuage_sandbox` is very useful fixture that we want to use in
multiple tests. It creates Nuage enterprise and its entities, and returns
its values, which can then be compared to ui and database values.
We moved it to `cfme/fixtures/nuage.py`, since its Nuage specific and
will only be used by Nuage tests.

`targeted_refresh` fixture checks logs, if targeted refresh was started
for specific targets. It can be used by other providers as well,
so it was moved to `cfme/fixtures/networks.py`.

Signed-off-by: sasoc <saso.cvitkovic@gmail.com>