Use log validator for targeted refresh
Targeted refesh fixture searches `evm.log` to check if refresh
was started for specific entity. Until this commit merkyl
plugin was used to tail `evm.log`, now LogValidator is used
to tail it. This change was made because LogValidator is
prefered over merkyl as ssh access is all you need for it.
While merkyl needs daemon to run on appliance and its failure
can affect targeted refresh test.

Upon migrating to the new approach, we had to enhance
LogValidator a bit to support our usecase. It's that until
now it was always used to search logs exactly once
(and fail test if not found), while we needed to periodically
retry the thing and only fail test if not fulfilled by the end of timeout.

Signed-off-by: sasoc <saso.cvitkovic@gmail.com>