Validate Network Router inventory
With this commit we create Nuage network router and
Nuage network router collection, which is then used to
validate its inventory.

We create sandbox with Nuage enterprise and its entities,
including network router. We use this sandbox to gather all
the stats we expect to be shown on the details page.
Then we find and navigate to created network router where
we compare its UI stats with expected ones.

We also use different steps to navigate to details page.
We go through provider using steps like this:
```
Nertwork -> Providers -> Provider (by name) -> Network Router -> Router (by name)
```
By using this steps we make sure that we are on the details
page of Nuage network router.

Signed-off-by: sasoc <saso.cvitkovic@gmail.com>