Add Redfish physical rack inventory validation
We extend existing physical rack functionality to enable
tests for inventory of Redfish phsyical racks. The physical
racks only shows the rack name as an inventory property
that we can check, but this also checks the correct
navigation to each of the Redfish racks discovered by
the provider.

Related Bugzilla tickets:
  * https://bugzilla.redhat.com/show_bug.cgi?id=1571610
  * https://bugzilla.redhat.com/show_bug.cgi?id=1574809