Validate Nuage network port inventory
With this commit we create Nuage network port and Nuage network
port collection, which we then use to validate details page of Nuage
network port.

We create sandbox with Nuage enterprise and its entities, including
network port. We use this sandbox to gather all the stats we expect
to be shown on the details page. Then we find and navigate to
created network port where we compare its UI stats with expected ones.

We also use different steps to navigate to details page. We go through
provider using steps like this:
```
Nertwork -> Providers -> Provider (by name) -> Network Ports -> Network Port (by name)
```
By using this steps we make sure that we are on the details page of
Nuage network port.

Signed-off-by: sasoc <saso.cvitkovic@gmail.com>