Expand Nuage sandbox
With this commit we expand Nuage sandbox with shared network
resources (cloud network) and floating ip.  We create a small
cloud network and use it to create floating ip, which is then associated
with vm vport. We need both of those entities for future
testing (e.g. inventory validation)

Sandbox fixture deploys Nuage enterprise (cloud tenant) and its entities.
We can use it to get values of those entities.

Signed-off-by: sasoc <saso.cvitkovic@gmail.com>