[1LP][WIP] Add a test for redfish physical server details (#7726)
* Add test for redfish physical server details

This test validates the statistics of a selected physical server
that has been inventorised by the Redfish provider as it is shown
in the GUI. It checks the displayed values against the ones
accessible through the Redfish API client.

* Include ems_ref in PhysicalServer

Extends the `PhysicalServer` class to include the database field
`ems_ref`. For some types of physical servers (e.g., in Redfish)
this represents the handle to the related resource in the API
server.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>

* Fix an expected phys. infra. adding title

This changes the expected title in the form for adding new physical
infrastructure providers. The error was unearthed after the
https://github.com/ManageIQ/integration_tests/commit/27c7ba9f8ee8797812c16774f3111357eda5472d#diff-8f65f83f3de1a967375f00c9f99ed6ccR514
commit.