Add test for validation of Nuage Cloud Tenant inventory
This test creates sandbox with Nuage cloud tenant and its entities.
We then navigate to Cloud tenant (which we created in previous step)
through provider, where we compare UI stats with sandbox stats.
We also extended fixture for nuage sandbox and made it 'module' scoped
which will help us with the future inventory tests.