[RFR] Integration tests for Redfish Physical Chassis and Racks (#7929)
* Add checking for Redfish Physical Chassis, Racks

For a Redfish Physical Infrastructure System, we want to
check the number of Physical Chassis and Physical Racks
in the Details page.

Related Bugzilla ticket:
  * https://bugzilla.redhat.com/show_bug.cgi?id=1571610

* Fix bug in PhysicalRack view

PhysicalRackEntity doesn't seem to have a pick method.

* Add Redfish physical rack inventory validation

We extend existing physical rack functionality to enable
tests for inventory of Redfish phsyical racks. The physical
racks only shows the rack name as an inventory property
that we can check, but this also checks the correct
navigation to each of the Redfish racks discovered by
the provider.

Related Bugzilla tickets:
  * https://bugzilla.redhat.com/show_bug.cgi?id=1571610
  * https://bugzilla.redhat.com/show_bug.cgi?id=1574809

* Add implementation of Physical Chassis widgets

We added the implementation of the elements in the GUI representing
and showing physical infrastructures' physical chassis.

* Implement testing of Redfish Physical Chassis

We added testing of the Redfish Physical Chassis GUI. We test
for chassis name, description and current status of the chassis
identifying LED, and we check the count of physical systems
enclosed within the chassis.