Test Nuage event playbook triggers
We copy AE instance (nuage_enterprise_create) on path `ManageIQ/System/Event/EmsEvent/Nuage`
and copy it to newly created domain. Then we create new AE method and attach it to previously
copied AE instance. This method is executed when Nuage Enterprise is created. Method runs
`printout.yaml` ansible playbook and display information about Nuage fields.
With the help of `LogValidator` we then check for appropriate line to appear in the `evm.log`.