Retry ssh connection if failed.
When ssh connection to additional appliances (for example in the update
tests) is being made, it could fail, because SSH was not fully up and
running.

This patch adds re-trying to create the ssh connections for 2 minutes.

The reraise was used to log-and-ignore the exceptions when running an
SSH command, but this looks to me as a bad practice as it just burries
the problems in the logs.