Add wait_for when checking the login page.
After the db migration and evm restart of the
   cfme/tests/test_db_migrate.py::test_db_migrate_replication[57]
and other tests with similar pattern, the login page is not fully loaded
immediately and needs some time to display all the components. If not
waited for, the test dies with: Exception: Could not open the login screen.

Adding simple wait_for helps to wait until the login page load.