[1LP][RFR] Fix minor version update tests. (#8521)
* Fix minor version update tests.

The if branch didn't make much sense. We definitely want to test
the z versions of 5.9.z as well as 5.10.z and definitely including updating
from z == 0

We need to test only the updates from x.y.z-1 to x.y.z versions. Fixing
the logic of testgen and making it more clear.

Increase timeouts because they seemed too short.

Wait for ssh availability instead of failing fast. SSH becomes
connective after a while.

Change the minor_build to z_version in test_appliance_update

* Fix test_update_replicated_webui and add single()

In replicated environment, we can have two masters. In such case one of
them (possibly incorrect one) was selected by the tests. This patch
tries to select the the correct rest-api object matching the appliance
in the context.

* Fix missing wait_for_evm_service.

It was replaced with evmserverd.wait_for_running

* Use browser stealing.

Hopefully it will mitigate

        try:
            return view.entities.get_entity(name=self.name, surf_pages=True, use_search=use_search)
        except ItemNotFound:
>           raise VmOrInstanceNotFound("VM '{}' not found in UI!".format(self.name))
E           VmOrInstanceNotFound: VM 'test_rest_db_U0K04Xe8Tk' not found in UI!

cfme/common/vm.py:240: VmOrInstanceNotFound

* Convert ha_appliances_with_providers to expect.

* Fix test_update_distributed_webui

* Add waits to update tests.

Add wait_for_web_ui and evmserverd.wait_for_running to prevent various
erros when connecting to the created and updated appliances.

* Fix failure trying deleting nonexisting VM.

The error
E           VmOrInstanceNotFound: VM 'test_rest_db_IZnmNTAi6N' not found in UI!
was raised when VM clone operation failed in the setup phase. This made
the tear-down failure.

* Provide more information on vmdb_version ssh command failure

* Unificate is_ha_monitor_started, fix it for cfme 5.10

* Fix ha_appliances_with_providers doc.

* Fix test_update_ha_webui, consolidate failover state getters

* Add safety check for the cfme version in the repo file.

The repo file (for example
http://file.cloudforms.lab.eng.rdu2.redhat.com/builds/cfme/5.10/5.10.3.1/packages.repo
) contains link to baseurl with dir `latest`. But not every time we have
picked the latest package.repo. In such case the package refered in
`latest` dir will not be the intended one and it is better to detect
that. This patch adds a method for checking the version of cfme the
packages.repo in the end ponts at.

* add preupdate version parameter

* Add wait_for when checking the login page.

After the db migration and evm restart of the
   cfme/tests/test_db_migrate.py::test_db_migrate_replication[57]
and other tests with similar pattern, the login page is not fully loaded
immediately and needs some time to display all the components. If not
waited for, the test dies with: Exception: Could not open the login screen.

Adding simple wait_for helps to wait until the login page load.

* fix replicated_appliances_with_providers

* Fix code formatting test_update_distributed_webui

* Fix test_update_embedded_ansible_webui

* Add BZ blocker: 1704835

* Add Accept header when waiting for evmserverd.

* Wait for failover-monitor to stabilize after restart

This should help with test_update_ha_webui

* Make flake8 happy

* Use LogValidator for inspecting the logs.

* Change waiting for appliance update in test_update_distributed_webui

* Fix error message formatting in get_apps

* Beautifulize the pytest_generate_tests

* Allow non-strict ssh host key checking.

* Fix 'ApplianceDB' object has no attribute 'postgres_version'

This is fixing

    @property
    def is_dedicated_active(self):
        result = self.appliance.ssh_client.run_command(
            "systemctl status {}-postgresql.service | grep running".format(
>               self.postgres_version))
E       AttributeError: 'ApplianceDB' object has no attribute 'postgres_version'

* Remove single() in favor of unpacking.

* Use six for the problematic imports

* Fix 404 when baseurl doesn't end with slash.

* Increase the timeout for update to make test_update_webui pass

* Add docblock about two masters problem to get_master

* Add condition dealing with BZ 1732092.

* Add yum strategy HA upgrade testing to the existing webui option

* Decrease the verbosity of ssh expect interacitions.